wip[gpu]: for kernel and framework for running pipeline
