+++ vortex-gpu/src/lib.rs
@@ -1,8 +1,10 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
-mod bit_unpack;
+pub mod bit_unpack;
+pub mod for_;
 mod take;
+mod task;
 
 pub use bit_unpack::cuda_bit_unpack;
 pub use take::cuda_take;
